Backport implement loading DDS textures at run-time
