+++ core/image.h
@@ -134,6 +134,7 @@ public:
 	static ImageMemLoadFunc _webp_mem_loader_func;
 	static ImageMemLoadFunc _tga_mem_loader_func;
 	static ImageMemLoadFunc _bmp_mem_loader_func;
+	static ImageMemLoadFunc _dds_mem_loader_func;
 
 	static void (*_image_compress_bc_func)(Image *, float, CompressSource p_source);
 	static void (*_image_compress_bptc_func)(Image *, float p_lossy_quality, CompressSource p_source);
@@ -343,6 +344,7 @@ public:
 	Error load_webp_from_buffer(const PoolVector<uint8_t> &p_array);
 	Error load_tga_from_buffer(const PoolVector<uint8_t> &p_array);
 	Error load_bmp_from_buffer(const PoolVector<uint8_t> &p_array);
+	Error load_dds_from_buffer(const PoolVector<uint8_t> &p_array);
 
 	Image(const uint8_t *p_mem_png_jpg, int p_len = -1);
 	Image(const char **p_xpm);
